+++ src/main/java/teammates/ui/webapi/output/SearchLinksStudentData.java
@@ -3,74 +3,20 @@ package teammates.ui.webapi.output;
 /**
  * Contains the links for a student.
  */
-public class SearchLinksStudentData {
-    protected boolean showLinks;
-    private String email;
-    private String manageAccountLink;
-    private String homePageLink;
-    private String courseJoinLink;
+public class SearchLinksStudentData extends CommonSearchLinksData {
     private String recordsPageLink;
 
     public SearchLinksStudentData() {
-        this.email = null;
-        this.manageAccountLink = null;
-        this.homePageLink = null;
-        this.courseJoinLink = null;
+        super();
         this.recordsPageLink = null;
-    }
-
-    public SearchLinksStudentData(
-            String email,
-            String manageAccountLink,
-            String homePageLink,
-            String courseJoinLink,
-            String recordsPageLink
-    ) {
-        this.email = email;
-        this.manageAccountLink = manageAccountLink;
-        this.homePageLink = homePageLink;
-        this.courseJoinLink = courseJoinLink;
-        this.recordsPageLink = recordsPageLink;
-    }
-
-    public void setEmail(String email) {
-        this.email = email;
-    }
-
-    public void setManageAccountLink(String manageAccountLink) {
-        this.manageAccountLink = manageAccountLink;
-    }
-
-    public void setHomePageLink(String homePageLink) {
-        this.homePageLink = homePageLink;
-    }
-
-    public void setCourseJoinLink(String courseJoinLink) {
-        this.courseJoinLink = courseJoinLink;
-    }
-
-    public void setRecordsPageLink(String recordsPageLink) {
-        this.recordsPageLink = recordsPageLink;
-    }
-
-    public String getEmail() {
-        return email;
-    }
-
-    public String getManageAccountLink() {
-        return manageAccountLink;
-    }
-
-    public String getHomePageLink() {
-        return homePageLink;
-    }
-
-    public String getCourseJoinLink() {
-        return courseJoinLink;
     }
 
     public String getRecordsPageLink() {
         return recordsPageLink;
+    }
+
+    public void setRecordsPageLink(String recordsPageLink) {
+        this.recordsPageLink = recordsPageLink;
     }
 }
